Moved _line into its own file, ready for optimizing

+++ src/lib/libgraph/moveto.c
@@ -1,7 +1,7 @@
 
 #include "_graph.h"
 
-void _moveto(short x, short y)
+void _moveto(int x, int y)
 {
     _graph.curx = x;
     _graph.cury = y;
